Fixed typos and extra warnings.

+++ test/PDS_find_first.c
@@ -13,8 +13,7 @@ int main()
         test_data(           "Klatu verata ni-mumble mumble", -1, '*', PDS_INVALID_VALUE ),
     end_test_data()
 
-
-    int i;
+    size_t i;
     test_foreach(i)
     {
         const char *first = test_str(i);
